feat: ship 0.8 with u64 docs, safe headers, and ingest/search speedups
Optional mimalloc and LE header I/O fix Issues #1/#2; NFS2/NWA2 plus analyzer,
BM25/phrase, durable flush, and write-path optimizations keep results correct
while improving exact and fuzzy query throughput.

+++ nanofts/examples/arrow_real_world.rs
@@ -80,8 +80,8 @@ fn main() {
         })
         .collect();
     
-    let all_doc_ids: Vec<Vec<u32>> = sizes.iter()
-        .map(|&size| (1..=size as u32).collect())
+    let all_doc_ids: Vec<Vec<u64>> = sizes.iter()
+        .map(|&size| (1..=size as u64).collect())
         .collect();
     
     // Simulate Arrow arrays
